colab_votacao.py
import google.auth
from google.auth.transport.requests import Request


def autenticar_usuario(creds):
    """Autentica e atualiza as credenciais, se necessário."""
    if creds and creds.expired and creds.refresh_token:
        creds.refresh(Request())
    return creds


class Votacao:
    """Classe para gerenciar e processar votos em uma eleição."""

    def __init__(self, candidatos):
        self.candidatos = candidatos

    def processar_voto(self, voto):
        """Processa o voto de um candidato."""
        if voto in self.candidatos:
            self.candidatos[voto] += 1
        else:
            print("Candidato inválido!")

    def exibir_resultados(self):
        """Exibe o resultado final da votação."""
        for candidato, votos em self.candidatos.items():
            print(f"{candidato}: {votos} votos")


def main():
    """Função principal que simula o processo de votação."""
    candidatos = {"Candidato A": 0, "Candidato B": 0}
    votacao = Votacao(candidatos)
    votos = ["Candidato A", "Candidato B", "Candidato A"]

    for voto in votos:
        votacao.processar_voto(voto)

    votacao.exibir_resultados()


if __name__ == "__main__":
    main()